backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
paho-mqtt==2.1.0
pydantic==2.9.2
pydantic-settings==2.6.0
pyyaml==6.0.2
python-dotenv==1.0.1
websockets==13.1
jinja2==3.1.4
pytest==8.3.4
httpx==0.27.2